dev-requirements.txt
flake8
kinto-client
mock
pytest
pytest-cache
pytest-capturelog
pytest-cover
pytest-sugar
tox
webtest
